Suggestion to approach passing arguments with macros differently.
Since Rust macros allow to define custom syntax, we can use syntax
that's already used in CLI programs: arguments are prepended with
hyphens. This hyphen syntax allows using aliases for arguments
(e.g. -f, --flag) without the need to make a Name struct and just
using 1 hyphen for one-character arguments, and 2 hyphens for
long-name arguments. Also, in CLI arguments a hyphen is used
instead of an underscore to separate words in an argument's name,
so here the macro can also replace the underscores in long-named
arguments with a hyphen (doesn't replace the prepended hyphens).

There are cases where a programmer wants to just get the value of
an argument or check if a flag is passed without saving the value
to a variable, so we can have the macro just get the value as an
expression and if they want to make a variable with the argument's
name they can write let before the argument: `flag!(let --my_arg)`.

Also added a few tweaks to make it more useful for a CLI program:
 - The macro skips the 0th argument which is always the exec file.
 - One-character flags can look in a group of flags (-abcd).
 - Replacing underscore with hyphens in long-named arguments.